+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Expect active development and potentially significant breaking changes in the `0.x` track. We'll try to be diligent about releasing a `1.0` version in a timely fashion (ideally within 3 to 6 months), to signal the start of a more stable API.
 
 ### vNEXT
-
+- Fixed the getFromAST functions so that they work with queries that have named fragments (i.e. working toward named fragment support)
 - Added basic guards to our Redux Store `subscribe` to prevent `broadcastQueries` from being called unnecessarily
 
 ### v0.3.12